==> 3_calculate_return_values.ipynb <==
# Calculate Return Values for a 1-in-10-year Event
---
**Project**: Masters Project <br>
**Author**: Nabig Chaudhry

# import necessary packages

import requests 
import numpy as np
import pandas as pd
import xarray as xr
from datetime import datetime
import os as os
from scipy import stats
import statsmodels.api as sm
import time

import lmoments3 as lm
from lmoments3 import distr as ldistr

import matplotlib.pyplot as plt
%matplotlib inline
import cartopy.crs as ccrs
import cartopy.feature as cfeature
import geoviews as gv
import holoviews as hv
import hvplot.pandas
import hvplot.xarray

# import functions specifically for masters project

%load_ext autoreload
%autoreload 2

import masters_project_functions as mp

## Step 1: Pull Intermediate Processed Data

os.listdir('./data/intermediate_processed')

import_path = './data/intermediate_processed'

cesm2_hist_1980 = xr.open_dataarray(f'{import_path}/cesm2_hist_1980.nc')
cesm2_ssp370_2020 = xr.open_dataarray(f'{import_path}/cesm2_ssp370_2020.nc')
cesm2_ssp370_2040 = xr.open_dataarray(f'{import_path}/cesm2_ssp370_2040.nc')
cesm2_ssp370_2060 = xr.open_dataarray(f'{import_path}/cesm2_ssp370_2060.nc')
cesm2_ssp370_2080 = xr.open_dataarray(f'{import_path}/cesm2_ssp370_2080.nc')

cnrm_hist_1980 = xr.open_dataarray(f'{import_path}/cnrm_hist_1980.nc')
cnrm_ssp370_2020 = xr.open_dataarray(f'{import_path}/cnrm_ssp370_2020.nc')
cnrm_ssp370_2040 = xr.open_dataarray(f'{import_path}/cnrm_ssp370_2040.nc')
cnrm_ssp370_2060 = xr.open_dataarray(f'{import_path}/cnrm_ssp370_2060.nc')
cnrm_ssp370_2080 = xr.open_dataarray(f'{import_path}/cnrm_ssp370_2080.nc')

## Step 2: Calculate Return Values for CESM2 Global Climate Model

cesm2_hist_1980_rv = mp.get_return_value(cesm2_hist_1980, 
                                         return_period=10, 
                                         bootstrap_runs=100)

cesm2_ssp370_2020_rv = mp.get_return_value(cesm2_ssp370_2020, 
                                           return_period=10, 
                                           bootstrap_runs=100)

cesm2_ssp370_2040_rv = mp.get_return_value(cesm2_ssp370_2040, 
                                           return_period=10, 
                                           bootstrap_runs=100)

cesm2_ssp370_2060_rv = mp.get_return_value(cesm2_ssp370_2060, 
                                           return_period=10, 
                                           bootstrap_runs=100)

cesm2_ssp370_2080_rv = mp.get_return_value(cesm2_ssp370_2080, 
                                           return_period=10, 
                                           bootstrap_runs=100)

## Step 3: Calculate Return Values for CNRM-ESM2-1 Global Climate Model

cnrm_hist_1980_rv = mp.get_return_value(cnrm_hist_1980, 
                                        return_period=10, 
                                        bootstrap_runs=100)

cnrm_ssp370_2020_rv = mp.get_return_value(cnrm_ssp370_2020, 
                                          return_period=10, 
                                          bootstrap_runs=100)

cnrm_ssp370_2040_rv = mp.get_return_value(cnrm_ssp370_2040, 
                                          return_period=10, 
                                          bootstrap_runs=100)

cnrm_ssp370_2060_rv = mp.get_return_value(cnrm_ssp370_2060, 
                                          return_period=10, 
                                          bootstrap_runs=100)

cnrm_ssp370_2080_rv = mp.get_return_value(cnrm_ssp370_2080, 
                                          return_period=10, 
                                          bootstrap_runs=100)

## Step 4: Export Final Data For Analysis and Visualization

export_path ='./data/final_for_analysis/return_value'

# export to data/final_for_analysis folder

cesm2_hist_1980_rv.to_netcdf(f'{export_path}/cesm2_hist_1980_rv.nc')
cesm2_ssp370_2020_rv.to_netcdf(f'{export_path}/cesm2_ssp370_2020_rv.nc')
cesm2_ssp370_2040_rv.to_netcdf(f'{export_path}/cesm2_ssp370_2040_rv.nc')
cesm2_ssp370_2060_rv.to_netcdf(f'{export_path}/cesm2_ssp370_2060_rv.nc')
cesm2_ssp370_2080_rv.to_netcdf(f'{export_path}/cesm2_ssp370_2080_rv.nc')

cnrm_hist_1980_rv.to_netcdf(f'{export_path}/cnrm_hist_1980_rv.nc')
cnrm_ssp370_2020_rv.to_netcdf(f'{export_path}/cnrm_ssp370_2020_rv.nc')
cnrm_ssp370_2040_rv.to_netcdf(f'{export_path}/cnrm_ssp370_2040_rv.nc')
cnrm_ssp370_2060_rv.to_netcdf(f'{export_path}/cnrm_ssp370_2060_rv.nc')
cnrm_ssp370_2080_rv.to_netcdf(f'{export_path}/cnrm_ssp370_2080_rv.nc')
