==> 6_visualize_trends_return_values.ipynb <==
# Visualize Trends in Return Values for a 1-in-10-year Event
---
**Project**: Masters Project <br>
**Author**: Nabig Chaudhry

# import necessary packages

import requests 
import numpy as np
import pandas as pd
import xarray as xr
from datetime import datetime
import os as os
from scipy import stats
import statsmodels.api as sm

import lmoments3 as lm
from lmoments3 import distr as ldistr

import matplotlib.pyplot as plt
from matplotlib import colors
%matplotlib inline
import cartopy.crs as ccrs
import cartopy.feature as cfeature
import geoviews as gv
import holoviews as hv
import hvplot.pandas
import hvplot.xarray
from bokeh.io import export_svgs

# import functions specifically for masters_project

%load_ext autoreload
%autoreload 2

import masters_project_functions as mp

## Step 1: Pull Final Return Value Data For Analysis

os.listdir('./data/final_for_analysis/return_value')

import_path = './data/final_for_analysis/return_value'

cesm2_hist_1980_rv = xr.open_dataset(f'{import_path}/cesm2_hist_1980_rv.nc')
cesm2_ssp370_2020_rv = xr.open_dataset(f'{import_path}/cesm2_ssp370_2020_rv.nc')
cesm2_ssp370_2040_rv = xr.open_dataset(f'{import_path}/cesm2_ssp370_2040_rv.nc')
cesm2_ssp370_2060_rv = xr.open_dataset(f'{import_path}/cesm2_ssp370_2060_rv.nc')
cesm2_ssp370_2080_rv = xr.open_dataset(f'{import_path}/cesm2_ssp370_2080_rv.nc')

cnrm_hist_1980_rv = xr.open_dataset(f'{import_path}/cnrm_hist_1980_rv.nc')
cnrm_ssp370_2020_rv = xr.open_dataset(f'{import_path}/cnrm_ssp370_2020_rv.nc')
cnrm_ssp370_2040_rv = xr.open_dataset(f'{import_path}/cnrm_ssp370_2040_rv.nc')
cnrm_ssp370_2060_rv = xr.open_dataset(f'{import_path}/cnrm_ssp370_2060_rv.nc')
cnrm_ssp370_2080_rv = xr.open_dataset(f'{import_path}/cnrm_ssp370_2080_rv.nc')

## Step 2: Pull Return Value Arrays for CESM2 Global Climate Model

cesm2_hist_1980_rv_array = mp.get_flat_array(cesm2_hist_1980_rv, 
                                             data_variable='return_value')

cesm2_hist_1980_rv_lower_array = mp.get_flat_array(cesm2_hist_1980_rv, 
                                             data_variable='conf_int_lower_limit')

cesm2_hist_1980_rv_upper_array = mp.get_flat_array(cesm2_hist_1980_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cesm2_hist_1980_df = pd.DataFrame({'return_value': cesm2_hist_1980_rv_array, 
                                   'return_value_lower': cesm2_hist_1980_rv_lower_array,
                                   'return_value_upper': cesm2_hist_1980_rv_upper_array})

cesm2_ssp370_2020_rv_array = mp.get_flat_array(cesm2_ssp370_2020_rv, 
                                             data_variable='return_value')

cesm2_ssp370_2020_rv_lower_array = mp.get_flat_array(cesm2_ssp370_2020_rv, 
                                             data_variable='conf_int_lower_limit')

cesm2_ssp370_2020_rv_upper_array = mp.get_flat_array(cesm2_ssp370_2020_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cesm2_ssp370_2020_df = pd.DataFrame({'return_value': cesm2_ssp370_2020_rv_array, 
                                     'return_value_lower': cesm2_ssp370_2020_rv_lower_array,
                                     'return_value_upper': cesm2_ssp370_2020_rv_upper_array})

cesm2_ssp370_2040_rv_array = mp.get_flat_array(cesm2_ssp370_2040_rv, 
                                             data_variable='return_value')

cesm2_ssp370_2040_rv_lower_array = mp.get_flat_array(cesm2_ssp370_2040_rv, 
                                             data_variable='conf_int_lower_limit')

cesm2_ssp370_2040_rv_upper_array = mp.get_flat_array(cesm2_ssp370_2040_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cesm2_ssp370_2040_df = pd.DataFrame({'return_value': cesm2_ssp370_2040_rv_array, 
                                     'return_value_lower': cesm2_ssp370_2040_rv_lower_array,
                                     'return_value_upper': cesm2_ssp370_2040_rv_upper_array})

cesm2_ssp370_2060_rv_array = mp.get_flat_array(cesm2_ssp370_2060_rv, 
                                             data_variable='return_value')

cesm2_ssp370_2060_rv_lower_array = mp.get_flat_array(cesm2_ssp370_2060_rv, 
                                             data_variable='conf_int_lower_limit')

cesm2_ssp370_2060_rv_upper_array = mp.get_flat_array(cesm2_ssp370_2060_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cesm2_ssp370_2060_df = pd.DataFrame({'return_value': cesm2_ssp370_2060_rv_array, 
                                     'return_value_lower': cesm2_ssp370_2060_rv_lower_array,
                                     'return_value_upper': cesm2_ssp370_2060_rv_upper_array})

cesm2_ssp370_2080_rv_array = mp.get_flat_array(cesm2_ssp370_2080_rv, 
                                             data_variable='return_value')

cesm2_ssp370_2080_rv_lower_array = mp.get_flat_array(cesm2_ssp370_2080_rv, 
                                             data_variable='conf_int_lower_limit')

cesm2_ssp370_2080_rv_upper_array = mp.get_flat_array(cesm2_ssp370_2080_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cesm2_ssp370_2080_df = pd.DataFrame({'return_value': cesm2_ssp370_2080_rv_array, 
                                     'return_value_lower': cesm2_ssp370_2080_rv_lower_array,
                                     'return_value_upper': cesm2_ssp370_2080_rv_upper_array})

# create dataframe from all return value arrays

cesm2_rv_df = pd.DataFrame({'cesm2_hist_1980': cesm2_hist_1980_rv_array, 
                            'cesm2_ssp370_2020': cesm2_ssp370_2020_rv_array,
                            'cesm2_ssp370_2040': cesm2_ssp370_2040_rv_array,
                            'cesm2_ssp370_2060': cesm2_ssp370_2060_rv_array,
                            'cesm2_ssp370_2080': cesm2_ssp370_2080_rv_array})

## Step 3: Pull Return Value Arrays for CNRM-ESM2-1 Global Climate Model

cnrm_hist_1980_rv_array = mp.get_flat_array(cnrm_hist_1980_rv, 
                                             data_variable='return_value')

cnrm_hist_1980_rv_lower_array = mp.get_flat_array(cnrm_hist_1980_rv, 
                                             data_variable='conf_int_lower_limit')

cnrm_hist_1980_rv_upper_array = mp.get_flat_array(cnrm_hist_1980_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cnrm_hist_1980_df = pd.DataFrame({'return_value': cnrm_hist_1980_rv_array, 
                                  'return_value_lower': cnrm_hist_1980_rv_lower_array,
                                  'return_value_upper': cnrm_hist_1980_rv_upper_array})

cnrm_ssp370_2020_rv_array = mp.get_flat_array(cnrm_ssp370_2020_rv, 
                                             data_variable='return_value')

cnrm_ssp370_2020_rv_lower_array = mp.get_flat_array(cnrm_ssp370_2020_rv, 
                                             data_variable='conf_int_lower_limit')

cnrm_ssp370_2020_rv_upper_array = mp.get_flat_array(cnrm_ssp370_2020_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cnrm_ssp370_2020_df = pd.DataFrame({'return_value': cnrm_ssp370_2020_rv_array, 
                                    'return_value_lower': cnrm_ssp370_2020_rv_lower_array,
                                    'return_value_upper': cnrm_ssp370_2020_rv_upper_array})

cnrm_ssp370_2040_rv_array = mp.get_flat_array(cnrm_ssp370_2040_rv, 
                                             data_variable='return_value')

cnrm_ssp370_2040_rv_lower_array = mp.get_flat_array(cnrm_ssp370_2040_rv, 
                                             data_variable='conf_int_lower_limit')

cnrm_ssp370_2040_rv_upper_array = mp.get_flat_array(cnrm_ssp370_2040_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cnrm_ssp370_2040_df = pd.DataFrame({'return_value': cnrm_ssp370_2040_rv_array, 
                                    'return_value_lower': cnrm_ssp370_2040_rv_lower_array,
                                    'return_value_upper': cnrm_ssp370_2040_rv_upper_array})

cnrm_ssp370_2060_rv_array = mp.get_flat_array(cnrm_ssp370_2060_rv, 
                                             data_variable='return_value')

cnrm_ssp370_2060_rv_lower_array = mp.get_flat_array(cnrm_ssp370_2060_rv, 
                                             data_variable='conf_int_lower_limit')

cnrm_ssp370_2060_rv_upper_array = mp.get_flat_array(cnrm_ssp370_2060_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cnrm_ssp370_2060_df = pd.DataFrame({'return_value': cnrm_ssp370_2060_rv_array, 
                                    'return_value_lower': cnrm_ssp370_2060_rv_lower_array,
                                    'return_value_upper': cnrm_ssp370_2060_rv_upper_array})

cnrm_ssp370_2080_rv_array = mp.get_flat_array(cnrm_ssp370_2080_rv, 
                                             data_variable='return_value')

cnrm_ssp370_2080_rv_lower_array = mp.get_flat_array(cnrm_ssp370_2080_rv, 
                                             data_variable='conf_int_lower_limit')

cnrm_ssp370_2080_rv_upper_array = mp.get_flat_array(cnrm_ssp370_2080_rv, 
                                             data_variable='conf_int_upper_limit')

# create dataframe of return value and associated confidence interval

cnrm_ssp370_2080_df = pd.DataFrame({'return_value': cnrm_ssp370_2080_rv_array, 
                                    'return_value_lower': cnrm_ssp370_2080_rv_lower_array,
                                    'return_value_upper': cnrm_ssp370_2080_rv_upper_array})

# create dataframe from all return value arrays

cnrm_rv_df = pd.DataFrame({'cnrm_hist_1980': cnrm_hist_1980_rv_array, 
                            'cnrm_ssp370_2020': cnrm_ssp370_2020_rv_array,
                            'cnrm_ssp370_2040': cnrm_ssp370_2040_rv_array,
                            'cnrm_ssp370_2060': cnrm_ssp370_2060_rv_array,
                            'cnrm_ssp370_2080': cnrm_ssp370_2080_rv_array})

## Step 4: Set Graphing Parameters for Box Plots

box_fill_color = '#fda847'
box_hover_color = 'blue'
y_axis_min=10
y_axis_max=60
width=650
height=400

columns = ['Historical 1980-99',
          'SSP3-7.0 2020-39',
          'SSP3-7.0 2040-59',
          'SSP3-7.0 2060-79',
          'SSP3-7.0 2080-99']

group_label='Time Periods'
value_label='Return Value (Degrees C)'

## Step 5: Visualize Trends and Return Value Data Through Box Plots

df = cesm2_rv_df.rename(columns={'cesm2_hist_1980':'Historical 1980-99', 
                                'cesm2_ssp370_2020':'SSP3-7.0 2020-39', 
                                'cesm2_ssp370_2040':'SSP3-7.0 2040-59',
                                'cesm2_ssp370_2060':'SSP3-7.0 2060-79',
                                'cesm2_ssp370_2080':'SSP3-7.0 2080-99'})

title = 'CESM2 Return Values (Degrees C) for 1-in-10-year Event'

cesm2_box_plot = df.hvplot.box(y=columns,
                               ylim=(y_axis_min, y_axis_max),
                               width=width,
                               height=height,
                               title=title,
                               box_fill_color=box_fill_color,
                               box_hover_color=box_hover_color,
                               group_label=group_label,
                               value_label=value_label)
cesm2_box_plot

df = cnrm_rv_df.rename(columns={'cnrm_hist_1980':'Historical 1980-99', 
                                'cnrm_ssp370_2020':'SSP3-7.0 2020-39', 
                                'cnrm_ssp370_2040':'SSP3-7.0 2040-59',
                                'cnrm_ssp370_2060':'SSP3-7.0 2060-79',
                                'cnrm_ssp370_2080':'SSP3-7.0 2080-99'})

title = 'CNRM-ESM2-1 Return Values (Degrees C) for 1-in-10-year Event'

cnrm_box_plot = df.hvplot.box(y=columns,
                              ylim=(y_axis_min, y_axis_max),
                              width=width,
                              height=height,
                              title=title,
                              box_fill_color=box_fill_color,
                              box_hover_color=box_hover_color,
                              group_label=group_label,
                              value_label=value_label)
cnrm_box_plot

## Step 6: Set Graphing Parameters for Line Plots

color_1 = '#fede81'
color_2 = '#fda847'
color_3 = '#fc5d2e'
color_4 = '#ee3123'
color_5 = '#900026'

y_axis_min=10
y_axis_max=60
width=850
height=450

label_1 = 'Historical 1980-99'
label_2 = 'SSP3-7.0 2020-39'
label_3 = 'SSP3-7.0 2040-59'
label_4 = 'SSP3-7.0 2060-79'
label_5 = 'SSP3-7.0 2080-99'

x_label='Index of Least to Greatest Return Values'
y_label='Return Value (Degrees C)'

## Step 7: Visualize Trends and Return Value Data Through Line Plots

# sort return value dataframes from low to high

cesm2_hist_1980_sorted = cesm2_hist_1980_df.sort_values('return_value').reset_index(drop=True)
cesm2_ssp370_2020_sorted = cesm2_ssp370_2020_df.sort_values('return_value').reset_index(drop=True)
cesm2_ssp370_2040_sorted = cesm2_ssp370_2040_df.sort_values('return_value').reset_index(drop=True)
cesm2_ssp370_2060_sorted = cesm2_ssp370_2060_df.sort_values('return_value').reset_index(drop=True)
cesm2_ssp370_2080_sorted = cesm2_ssp370_2080_df.sort_values('return_value').reset_index(drop=True)

cnrm_hist_1980_sorted = cnrm_hist_1980_df.sort_values('return_value').reset_index(drop=True)
cnrm_ssp370_2020_sorted = cnrm_ssp370_2020_df.sort_values('return_value').reset_index(drop=True)
cnrm_ssp370_2040_sorted = cnrm_ssp370_2040_df.sort_values('return_value').reset_index(drop=True)
cnrm_ssp370_2060_sorted = cnrm_ssp370_2060_df.sort_values('return_value').reset_index(drop=True)
cnrm_ssp370_2080_sorted = cnrm_ssp370_2080_df.sort_values('return_value').reset_index(drop=True)

title = 'CESM2 Return Values (Degrees C) for 1-in-10-year Event'

hv_line = cesm2_hist_1980_sorted.hvplot.line(y='return_value',
                                              ylim=(y_axis_min, y_axis_max),
                                              width=width,
                                              height=height,
                                              title=title,
                                              xlabel=x_label, 
                                              ylabel=y_label,
                                              label=label_1).opts(color=color_1) * \
cesm2_ssp370_2020_sorted.hvplot.line(y='return_value',
                                    label=label_2).opts(color=color_2) * \
cesm2_ssp370_2040_sorted.hvplot.line(y='return_value',
                                    label=label_3).opts(color=color_3) * \
cesm2_ssp370_2060_sorted.hvplot.line(y='return_value',
                                    label=label_4).opts(color=color_4) * \
cesm2_ssp370_2080_sorted.hvplot.line(y='return_value',
                                    label=label_5).opts(color=color_5) * \
cesm2_hist_1980_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_2,
                                   alpha=0.15).opts(color=color_2) * \
cesm2_ssp370_2020_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_2,
                                   alpha=0.15).opts(color=color_2) * \
cesm2_ssp370_2040_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_3,
                                   alpha=0.15).opts(color=color_3) * \
cesm2_ssp370_2060_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_4,
                                   alpha=0.15).opts(color=color_4) * \
cesm2_ssp370_2080_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_5,
                                   alpha=0.15).opts(color=color_5)

cesm2_line_plot = hv_line.opts(legend_position='right')
cesm2_line_plot

title = 'CNRM-ESM2-1 Return Values (Degrees C) for 1-in-10-year Event'

hv_line = cnrm_hist_1980_sorted.hvplot.line(y='return_value',
                                              ylim=(y_axis_min, y_axis_max),
                                              width=width,
                                              height=height,
                                              title=title,
                                              xlabel=x_label, 
                                              ylabel=y_label,
                                              label=label_1).opts(color=color_1) * \
cnrm_ssp370_2020_sorted.hvplot.line(y='return_value',
                                    label=label_2).opts(color=color_2) * \
cnrm_ssp370_2040_sorted.hvplot.line(y='return_value',
                                    label=label_3).opts(color=color_3) * \
cnrm_ssp370_2060_sorted.hvplot.line(y='return_value',
                                    label=label_4).opts(color=color_4) * \
cnrm_ssp370_2080_sorted.hvplot.line(y='return_value',
                                    label=label_5).opts(color=color_5) * \
cnrm_hist_1980_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_2,
                                   alpha=0.15).opts(color=color_2) * \
cnrm_ssp370_2020_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_2,
                                   alpha=0.15).opts(color=color_2) * \
cnrm_ssp370_2040_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_3,
                                   alpha=0.15).opts(color=color_3) * \
cnrm_ssp370_2060_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_4,
                                   alpha=0.15).opts(color=color_4) * \
cnrm_ssp370_2080_sorted.hvplot.area(y='return_value_lower', 
                                   y2='return_value_upper',
                                   line_color=color_5,
                                   alpha=0.15).opts(color=color_5)

cnrm_line_plot = hv_line.opts(legend_position='right')
cnrm_line_plot

## Step 8: Export Return Value Trend Visualizations

export_int_image_path = './images/interactive/return_value'
export_stat_image_path = './images/static/return_value'

# export visualizations as interactive html images

hv.save(cesm2_line_plot, f'{export_int_image_path}/cesm2_line_plot.html')
hv.save(cnrm_line_plot, f'{export_int_image_path}/cnrm_line_plot.html')

# define function to export visualizations as static svg files

def export_svg(obj, filename):
    plot_state = hv.renderer('bokeh').get_plot(obj).state
    plot_state.output_backend = 'svg'
    export_svgs(plot_state, filename=filename)

# export visualizations as static svg images

export_svg(cesm2_box_plot, f'{export_stat_image_path}/cesm2_box_plot.svg')
export_svg(cnrm_box_plot, f'{export_stat_image_path}/cnrm_box_plot.svg')

export_svg(cesm2_line_plot, f'{export_stat_image_path}/cesm2_line_plot.svg')
export_svg(cnrm_line_plot, f'{export_stat_image_path}/cnrm_line_plot.svg')
